package reorg, add more class/method docs
