+++ src/main/kotlin/xyz/block/moneyaddress/MoneyAddress.kt
@@ -3,6 +3,11 @@ package xyz.block.moneyaddress
 import web5.sdk.dids.didcore.Service
 import xyz.block.moneyaddress.urn.DapUrn
 
+/**
+ * A generic representation of a MoneyAddress, without specific details of the currency or protocol.
+ *
+ * For a more specific representation, see [TypedMoneyAddress].
+ */
 data class MoneyAddress(
   val id: String,
   val urn: DapUrn,
@@ -19,6 +24,12 @@ data class MoneyAddress(
   }
 }
 
+/**
+ * Extracts the MoneyAddress objects from a DID Service object.
+ *
+ * @throws InvalidMoneyAddressException if the service type is not "MoneyAddress".
+ * @throws InvalidDapUrnException if the URN is not a valid DAP URN.
+ */
 fun Service.toMoneyAddresses(): List<MoneyAddress> {
   if (type != MoneyAddress.KIND) {
     throw InvalidMoneyAddressException
@@ -29,6 +40,9 @@ fun Service.toMoneyAddresses(): List<MoneyAddress> {
   }
 }
 
+/**
+ * Converts a DAP URN to a MoneyAddress.
+ */
 fun DapUrn.toMoneyAddress(id: String): MoneyAddress =
   MoneyAddress(
     id = id,
