+++ src/test/kotlin/xyz/block/moneyaddress/typed/BtcOnChainAddressTest.kt
@@ -1,7 +1,7 @@
-package xyz.block.moneyaddress.typed.x
+package xyz.block.moneyaddress.typed
 
-import xyz.block.moneyaddress.typed.MoneyAddressTest.Companion.btcOnChainMoneyAddress1
-import xyz.block.moneyaddress.typed.x.BtcOnChainAddress.Companion.toBtcOnChainAddress
+import xyz.block.moneyaddress.TypedMoneyAddressTest.Companion.btcOnChainMoneyAddress1
+import xyz.block.moneyaddress.typed.BtcOnChainAddress.Companion.toBtcOnChainAddress
 import kotlin.test.Test
 import kotlin.test.assertEquals
